requirements.txt
# Dependencias principales
pandas>=1.3.0
numpy>=1.21.0
ccxt>=2.0.0
redis>=4.0.0
pyyaml>=5.4.1
structlog>=21.1.0
prometheus-client>=0.11.0
pydantic>=1.8.2

# API y Autenticación
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=1.1.3

# Webhooks y Notificaciones
aiohttp>=3.8.0
backoff>=1.10.0
tenacity>=8.0.0

# Almacenamiento en la nube
boto3>=1.18.0
google-cloud-storage>=1.42.0
azure-storage-blob>=12.9.0

# Logging y monitoreo
python-json-logger>=2.0.0
colorama>=0.4.4
rich>=10.0.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-mock>=3.6.0
pytest-asyncio>=0.15.1

# Desarrollo
black>=21.6b0
isort>=5.9.2
flake8>=3.9.2
mypy>=0.910
pylint>=2.8.0

# Documentación
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
sphinx-autodoc-typehints>=1.12.0

# Utilidades
python-dateutil>=2.8.0
pytz>=2021.1
requests>=2.25.0
aiohttp>=3.8.0
backoff>=1.10.0
tenacity>=8.0.0

# Tipos
types-PyYAML>=6.0.0
types-redis>=4.0.0
types-requests>=2.25.0
types-python-dateutil>=2.8.0

# APIs y Data
python-binance==1.0.17
finnhub-python==2.4.18
polygon-api-client==1.12.3
twelvedata==1.2.11
alpha_vantage==2.3.1

# Análisis y Procesamiento
scipy>=1.7.0
scikit-learn>=0.24.2
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
streamlit>=1.12.0
statsmodels==0.14.0
ta-lib==0.4.28
shap>=0.41.0

# Visualización y UI
dash>=2.6.0
dash-bootstrap-components>=1.2.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
plotly-resampler>=0.8.0
kaleido>=0.2.0

# Infraestructura
python-dotenv>=0.19.0

# Logging y Monitoreo
prometheus-client==0.17.1

# Testing
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilidades
requests>=2.26.0
aiohttp==3.8.5
tqdm>=4.62.0
joblib>=1.0.0

# Additional dependencies
typing-extensions>=3.10.0
flake8>=3.9.0
mypy>=0.910

# Caching and serialization
msgpack>=1.0.2

# Monitoring and logging
psutil>=5.8.0

# Backtesting and Optimization
optuna>=3.0.0  # For parameter optimization
scikit-learn>=1.0.0  # For cross-validation
pandas-ta>=0.3.0  # Technical analysis indicators
empyrical>=0.5.5  # Performance metrics
vectorbt>=0.24.0  # Vectorized backtesting
pyfolio>=0.9.2  # Portfolio analysis

# Seguridad y Encriptación
cryptography>=3.4.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Rate Limiting y Cache
redis>=4.0.0
fastapi-limiter>=0.1.5

# Documentación API
fastapi-swagger-ui>=0.1.0 